Update README.md
He is the "Son of David", a "king", and the Messiah. Luke presents Jesus as the divine-human saviour who shows compassion to the needy. He is the friend of sinners and outcasts, and came to seek and save the lost. This gospel includes well-known parables, such as the Good Samaritan and the Prodigal Son